P9: Finalized Merged Fragments
