+++ app/build.gradle.kts
@@ -54,6 +54,7 @@ android {
 }
 
 dependencies {
+
     implementation("androidx.navigation:navigation-fragment-ktx:2.9.1")
     implementation("androidx.navigation:navigation-ui-ktx:2.9.1")
     implementation("com.google.android.material:material:1.12.0")
