+++ app/src/main/java/com/mobdeve/agbuya/hallar/hong/fridge/container/ContainerActivityFragmentMain.kt
@@ -1,5 +1,6 @@
 package com.mobdeve.agbuya.hallar.hong.fridge.container
 
+import android.app.Activity.RESULT_OK
 import android.content.Intent
 import android.os.Build
 import android.os.Build.VERSION.SDK_INT
@@ -8,17 +9,21 @@ import android.os.Parcelable
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import android.widget.FrameLayout
+import androidx.activity.result.ActivityResult
+import androidx.activity.result.contract.ActivityResultContracts
 import androidx.fragment.app.Fragment
 import androidx.navigation.fragment.findNavController
 import androidx.recyclerview.widget.LinearLayoutManager
 import com.mobdeve.agbuya.hallar.hong.fridge.R
-import com.mobdeve.agbuya.hallar.hong.fridge.adapter.ContainerActivityAdapter
-import com.mobdeve.agbuya.hallar.hong.fridge.adapter.ContainerActivityEditAdapter
+
+import com.mobdeve.agbuya.hallar.hong.fridge.adapter.ContainerActivityMainAdapter
+import com.mobdeve.agbuya.hallar.hong.fridge.databinding.BaseSearchbarContainerBinding
 import com.mobdeve.agbuya.hallar.hong.fridge.domain.ContainerModel
 import com.mobdeve.agbuya.hallar.hong.fridge.databinding.ContainerActivityMainBinding
 
 
-class ContainerActivityFragment : Fragment() {
+class ContainerActivityFragmentMain : Fragment() {
     companion object{
         val CONTAINERS_KEY : String = "CONTAINER_DATA_KEY"
 
@@ -26,9 +31,15 @@ class ContainerActivityFragment : Fragment() {
     private var _binding:ContainerActivityMainBinding? = null
     private val binding get() = _binding!!
 
-    private lateinit var postList:ArrayList<ContainerModel>
+    private lateinit var containerList:ArrayList<ContainerModel>
 
-
+    private val newContainerResultLauncher = registerForActivityResult(
+        ActivityResultContracts.StartActivityForResult()
+    ) { result: ActivityResult ->
+        if (result.resultCode == RESULT_OK) {
+            // handle returned result if needed
+        }
+    }
     // These two inlines suppresses deprecation errors
     inline fun <reified T : Parcelable> Intent.parcelableArrayList(key: String): ArrayList<T>? = when {
         SDK_INT >= 33 -> getParcelableArrayListExtra(key, T::class.java)
@@ -40,11 +51,11 @@ class ContainerActivityFragment : Fragment() {
     }
 
     fun loadFunnyData(){
-        postList= ContainerDataHelper.initializeContainers(requireContext())
+        containerList= ContainerDataHelper.initializeContainers(requireContext())
 
     }
-    fun getPostList():ArrayList<ContainerModel>{
-        return postList
+    fun getContainerList():ArrayList<ContainerModel>{
+        return containerList
     }
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
@@ -55,7 +66,7 @@ class ContainerActivityFragment : Fragment() {
 
         if(savedInstanceState!= null){
             if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.TIRAMISU) {
-                postList = savedInstanceState.getParcelableArrayList<ContainerModel>(CONTAINERS_KEY, ContainerModel::class.java) ?: run {
+                containerList = savedInstanceState.getParcelableArrayList<ContainerModel>(CONTAINERS_KEY, ContainerModel::class.java) ?: run {
                     ContainerDataHelper.initializeContainers(requireContext())
                 }
             }
@@ -66,12 +77,8 @@ class ContainerActivityFragment : Fragment() {
 
 
         binding.apply {
-            containerRecyclerView.adapter = ContainerActivityAdapter(postList){
-                findNavController().navigate(R.id.gotoContainerEdit)
-            }
-            containerRecyclerView.layoutManager = LinearLayoutManager(requireContext())
             addContainerBtn.setOnClickListener {
-                findNavController().navigate(R.id.gotoContainerEdit)
+
             }
         }
 
@@ -79,19 +86,49 @@ class ContainerActivityFragment : Fragment() {
         return view
     }
 
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+
+        setupTopBar()
+        setupRecycler()
+    }
     override fun onDestroyView() {
         super.onDestroyView()
         _binding = null
     }
     override fun onSaveInstanceState(outState: Bundle) {
         super.onSaveInstanceState(outState)
-        outState.putParcelableArrayList(CONTAINERS_KEY, postList)
+        outState.putParcelableArrayList(CONTAINERS_KEY, containerList)
 
     }
 
-//    private fun clickListener() = with(binding){
-//        goSecondFragment.setOnClickListener{
-//            findNavController().navigate(R.id.action_global_secondFragment)
-//        }
-//    }
+    private fun setupTopBar() {
+
+        binding.searchBarContainer.containerHeaderTv.setText(R.string.add_container)
+    }
+
+    private fun setupRecycler() {
+
+
+        binding.containerRecyclerView.apply {
+            layoutManager = LinearLayoutManager(requireContext())
+            adapter = ContainerActivityMainAdapter(containerList) {
+                val action = R.id.gotoContainerEdit
+                val bundle = Bundle().apply {
+                    putInt("EDIT_TYPE", EditType.EDIT.ordinal)
+                }
+                findNavController().navigate(action, bundle)
+            }
+        }
+
+        binding.addContainerBtn.setOnClickListener {
+            val action = R.id.gotoContainerEdit
+            val bundle = Bundle().apply {
+                putInt("EDIT_TYPE", EditType.ADD.ordinal)
+            }
+            findNavController().navigate(action, bundle)
+        }
+    }
+
 }
