+++ app/src/main/res/layout/activity_main.xml
@@ -1,43 +1,24 @@
 <?xml version="1.0" encoding="utf-8"?>
 <androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent">
 
-    <FrameLayout
-        android:id="@+id/topBar"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:elevation="10dp"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintVertical_bias="0.0">
 
-        <include
-            android:id="@+id/searchBarContainer"
-            layout="@layout/base_searchbar_container"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:elevation="4dp" />
-    </FrameLayout>
-
-    <FrameLayout
+    <androidx.fragment.app.FragmentContainerView
         android:id="@+id/dynamicContent"
+        android:name="androidx.navigation.fragment.NavHostFragment"
         android:layout_width="0dp"
         android:layout_height="0dp"
         android:layout_marginTop="-10dp"
         android:elevation="0dp"
+        app:defaultNavHost="true"
         app:layout_constraintBottom_toTopOf="@+id/navBarLayout"
-        app:layout_constraintEnd_toEndOf="@+id/topBar"
-        app:layout_constraintStart_toStartOf="@+id/topBar"
-        app:layout_constraintTop_toBottomOf="@+id/topBar"
-        app:layout_constraintVertical_bias="0.0">
-
-
-    </FrameLayout>
-
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="parent"
+        app:navGraph="@navigation/nav_graph" />
 
     <FrameLayout
         android:id="@+id/navBarLayout"
