+++ app/src/main/res/layout/container_activity_edit.xml
@@ -7,6 +7,24 @@
     android:layout_height="match_parent"
     tools:ignore="MergeRootFrame">
 
+    <FrameLayout
+        android:id="@+id/topBar"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:elevation="10dp"
+        app:layout_constraintBottom_toBottomOf="parent"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="parent"
+        app:layout_constraintVertical_bias="0.0">
+
+        <include
+            android:id="@+id/base_container_title_topBar"
+            layout="@layout/base_container"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:elevation="4dp" />
+    </FrameLayout>
 
     <LinearLayout
         android:id="@+id/linearLayout2"
@@ -19,7 +37,7 @@
         android:paddingEnd="100dp"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent">
+        app:layout_constraintTop_toBottomOf="@id/topBar">
 
         <TextView
             android:id="@+id/containerNameLabel_tv"
